Implement RFC 7539

+++ include/openssl/chacha.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014 - 2015, Kurt Cancemi (kurt@x64architecture.com)
+ * Copyright (c) 2014 - 2016, Kurt Cancemi (kurt@x64architecture.com)
  *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -44,7 +44,7 @@ void ChaCha(ChaCha_ctx *ctx, uint8_t *out, const uint8_t *in,
             size_t inlen);
 
 void CRYPTO_chacha_20(uint8_t *out, const uint8_t *in, size_t inlen,
-                      const uint8_t key[32], const uint8_t nonce[8], size_t counter);
+                      const uint8_t key[32], const uint8_t nonce[8], uint64_t counter);
 
 #ifdef __cplusplus
 }
